Modify queue message received structure

+++ notifications/src/main/java/com/akgym/notifications/listener/PaymentInfo.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 
 public record PaymentInfo(
     UUID id,
-    int timestamp
+    long timestamp
 ) implements Serializable {
 }
